# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Envelope_Summary_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/Summary"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[Summary]",
        "type": "object"
      },
...
      "Envelope_UsageDetail_": {
        "properties": {
          "data": {
            "$ref": "#/components/schemas/UsageDetail"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[UsageDetail]",
        "type": "object"
      },
...
      "Envelope_list_TimelinePoint__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/TimelinePoint"
            },
            "title": "Data",
            "type": "array"
          }
        },
        "required": [
          "data"
        ],
        "title": "Envelope[list[TimelinePoint]]",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "PageMeta": {
        "properties": {
          "page": {
            "title": "Page",
            "type": "integer"
          },
          "page_size": {
            "title": "Page Size",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "total",
          "page",
          "page_size"
        ],
        "title": "PageMeta",
        "type": "object"
      },
...
      "PagedEnvelope_list_UsageRow__": {
        "properties": {
          "data": {
            "items": {
              "$ref": "#/components/schemas/UsageRow"
            },
            "title": "Data",
            "type": "array"
          },
          "meta": {
            "$ref": "#/components/schemas/PageMeta"
          }
        },
        "required": [
          "data",
          "meta"
        ],
        "title": "PagedEnvelope[list[UsageRow]]",
        "type": "object"
      },
...
      "Summary": {
        "properties": {
          "avg_tokens_per_request": {
            "default": 0,
            "title": "Avg Tokens Per Request",
            "type": "number"
          },
          "cache_read_tokens": {
            "default": 0,
            "title": "Cache Read Tokens",
            "type": "integer"
          },
          "cache_write_tokens": {
            "default": 0,
            "title": "Cache Write Tokens",
            "type": "integer"
          },
          "input_tokens": {
            "default": 0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "output_tokens": {
            "default": 0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "top_client": {
            "default": "",
            "title": "Top Client",
            "type": "string"
          },
          "top_model": {
            "default": "",
            "title": "Top Model",
            "type": "string"
          },
          "total_projects": {
            "title": "Total Projects",
            "type": "integer"
          },
          "total_requests": {
            "title": "Total Requests",
            "type": "integer"
          },
          "total_sessions": {
            "title": "Total Sessions",
            "type": "integer"
          },
          "total_tokens": {
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "total_tokens",
          "total_requests",
          "total_projects",
          "total_sessions"
        ],
        "title": "Summary",
        "type": "object"
      },
      "TimelinePoint": {
        "properties": {
          "cache_read": {
            "title": "Cache Read",
            "type": "integer"
          },
          "cache_write": {
            "title": "Cache Write",
            "type": "integer"
          },
          "day": {
            "title": "Day",
            "type": "string"
          },
          "input": {
            "title": "Input",
            "type": "integer"
          },
          "output": {
            "title": "Output",
            "type": "integer"
          },
          "requests": {
            "default": 0,
            "title": "Requests",
            "type": "integer"
          },
          "tokens": {
            "title": "Tokens",
            "type": "integer"
          }
        },
        "required": [
          "day",
          "tokens",
          "input",
          "output",
          "cache_read",
          "cache_write"
        ],
        "title": "TimelinePoint",
        "type": "object"
      },
...
      "UsageDetail": {
        "properties": {
          "cache_read_tokens": {
            "default": 0,
            "title": "Cache Read Tokens",
            "type": "integer"
          },
          "cache_write_tokens": {
            "default": 0,
            "title": "Cache Write Tokens",
            "type": "integer"
          },
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client"
          },
          "context_window": {
            "default": 0,
            "title": "Context Window",
            "type": "integer"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "event_time": {
            "title": "Event Time",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "input_tokens": {
            "default": 0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "max_output_tokens": {
            "default": 0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "output_tokens": {
            "default": 0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "prompt_full": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "Full prompt context; null if full-text storage is off",
            "title": "Prompt Full"
          },
          "prompt_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Preview"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "response_full": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Full"
          },
          "response_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Preview"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          },
          "transcript_line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Line"
          },
          "transcript_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript Path"
          }
        },
        "required": [
          "id",
          "event_time",
          "session_id"
        ],
        "title": "UsageDetail",
        "type": "object"
      },
      "UsageRow": {
        "properties": {
          "cache_read_tokens": {
            "default": 0,
            "title": "Cache Read Tokens",
            "type": "integer"
          },
          "cache_write_tokens": {
            "default": 0,
            "title": "Cache Write Tokens",
            "type": "integer"
          },
          "client": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Client"
          },
          "context_window": {
            "default": 0,
            "title": "Context Window",
            "type": "integer"
          },
          "cwd": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cwd"
          },
          "event_time": {
            "title": "Event Time",
            "type": "string"
          },
          "id": {
            "title": "Id",
            "type": "integer"
          },
          "input_tokens": {
            "default": 0,
            "title": "Input Tokens",
            "type": "integer"
          },
          "max_output_tokens": {
            "default": 0,
            "title": "Max Output Tokens",
            "type": "integer"
          },
          "model": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model"
          },
          "output_tokens": {
            "default": 0,
            "title": "Output Tokens",
            "type": "integer"
          },
          "project": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Project"
          },
          "prompt_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Prompt Preview"
          },
          "provider": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Provider"
          },
          "response_preview": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Preview"
          },
          "session_id": {
            "title": "Session Id",
            "type": "string"
          },
          "total_tokens": {
            "default": 0,
            "title": "Total Tokens",
            "type": "integer"
          }
        },
        "required": [
          "id",
          "event_time",
          "session_id"
        ],
        "title": "UsageRow",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Read-only REST API over the local Token Telemetry database. Binds to 127.0.0.1; see docs/API.md for conventions (envelopes, filters, pagination, time zones).",
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "summary": "Local-first token, tool, skill and MCP observability for Claude Code.",
    "title": "Token Telemetry API",
    "version": "2.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/usage": {
      "get": {
        "operationId": "list_usage_api_v1_usage_get",
        "parameters": [
          {
            "in": "query",
            "name": "sort",
            "required": false,
            "schema": {
              "default": "time",
              "pattern": "^(time|tokens)$",
              "title": "Sort",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "order",
            "required": false,
            "schema": {
              "default": "desc",
              "pattern": "^(asc|desc)$",
              "title": "Order",
              "type": "string"
            }
          },
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope_list_UsageRow__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "List requests",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/project/{project}": {
      "get": {
        "operationId": "usage_by_project_api_v1_usage_project__project__get",
        "parameters": [
          {
            "in": "path",
            "name": "project",
            "required": true,
            "schema": {
              "title": "Project",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "minimum": 1,
              "title": "Page",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 100,
              "maximum": 1000,
              "minimum": 1,
              "title": "Page Size",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagedEnvelope_list_UsageRow__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Requests for one project",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/summary": {
      "get": {
        "operationId": "summary_api_v1_usage_summary_get",
        "parameters": [
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_Summary_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Totals for the filtered range",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/timeline": {
      "get": {
        "operationId": "timeline_api_v1_usage_timeline_get",
        "parameters": [
          {
            "description": "0 = full history, else the N most recent active days",
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 0,
              "description": "0 = full history, else the N most recent active days",
              "maximum": 3650,
              "minimum": 0,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "description": "Project name ('All' = no filter)",
            "in": "query",
            "name": "project",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Project name ('All' = no filter)",
              "title": "Project"
            }
          },
          {
            "description": "Client name ('All' = no filter)",
            "in": "query",
            "name": "client",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Client name ('All' = no filter)",
              "title": "Client"
            }
          },
          {
            "in": "query",
            "name": "model",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          },
          {
            "in": "query",
            "name": "session_id",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Session Id"
            }
          },
          {
            "description": "First day, YYYY-MM-DD (local)",
            "in": "query",
            "name": "start",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "First day, YYYY-MM-DD (local)",
              "title": "Start"
            }
          },
          {
            "description": "Last day, YYYY-MM-DD (local, inclusive)",
            "in": "query",
            "name": "end",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Last day, YYYY-MM-DD (local, inclusive)",
              "title": "End"
            }
          },
          {
            "description": "Browser getTimezoneOffset() in minutes",
            "in": "query",
            "name": "tz_offset",
            "required": false,
            "schema": {
              "default": 0,
              "description": "Browser getTimezoneOffset() in minutes",
              "maximum": 840,
              "minimum": -840,
              "title": "Tz Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_list_TimelinePoint__"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Daily totals (local days)",
        "tags": [
          "usage"
        ]
      }
    },
    "/api/v1/usage/{usage_id}": {
      "get": {
        "operationId": "usage_detail_api_v1_usage__usage_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "usage_id",
            "required": true,
            "schema": {
              "title": "Usage Id",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Envelope_UsageDetail_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "One request, with full text",
        "tags": [
          "usage"
        ]
      }
    },
...
  }
}